\name{FQLearn-package}
\alias{FQLearn-package}
\alias{FQLearn}
\docType{package}
\title{
\packageTitle{FQLearn}
}
\description{
\packageDescription{FQLearn}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FQLearn}
\packageIndices{FQLearn}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FQLearn}

Maintainer: \packageMaintainer{FQLearn}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
